+++ packages/extension/src/ui/features/accountNfts/NFT.tsx
@@ -4,6 +4,8 @@ import { Account } from "../accounts/Account"
 import { NftItem } from "./NftItem"
 import { useNFT } from "./useNfts"
 import { Flex, Spinner } from "@chakra-ui/react"
+import { useNavigate } from "react-router-dom"
+import { routes } from "../../routes"
 
 interface NftProps {
   collectionId: string
@@ -13,6 +15,7 @@ interface NftProps {
 }
 
 const Nft: FC<NftProps> = ({ collectionId, nftId, account, network }) => {
+  const navigate = useNavigate()
   const { nft } = useNFT(network, collectionId, nftId, account)
   if (nft === undefined) {
     return (
@@ -22,11 +25,9 @@ const Nft: FC<NftProps> = ({ collectionId, nftId, account, network }) => {
     )
   }
   return <NftItem
+    onClick={() => navigate(routes.sendNft(collectionId, nftId))}
     thumbnailSrc={nft.metadata.image || ""}
-    name={
-      nft.metadata.name ||
-      "Untitled"
-    }
+    name={nft.metadata.name || "Untitled"}
   />
 }
