+++ packages/extension/src/ui/features/accountNfts/NftFigure.tsx
@@ -2,11 +2,10 @@ import { Box } from "@chakra-ui/react"
 import { FC, ReactNode } from "react"
 
 interface NftFigureProps {
-  onClick: () => void
   children: ReactNode
 }
 
-const NftFigure: FC<NftFigureProps> = ({ onClick, children }) => (
+const NftFigure: FC<NftFigureProps> = ({ children }) => (
   <Box
     w="160px"
     h="192px"
@@ -19,7 +18,6 @@ const NftFigure: FC<NftFigureProps> = ({ onClick, children }) => (
     data-group
     borderRadius="lg"
     p="2"
-    onClick={onClick}
     _hover={{ backgroundColor: "neutrals.700" }}
   >
     {children}
